@@ -1,6 +1,6 @@
 {
   "name": "rezoning-web",
-  "version": "1.0.11",
+  "version": "1.2.0",
   "description": "Frontend application for the REZoning project",
   "repository": {
     "type": "git",
@@ -31,7 +31,7 @@
     "cy:run": "NODE_ENV=production cypress run"
   },
   "engines": {
-    "node": "12.x"
+    "node": "18.x"
   },
   "browserify": {
     "transform": [
@@ -98,8 +98,6 @@
     "shapefile": "^0.6.6",
     "simplify-geojson": "^1.0.3",
     "start-server-and-test": "^1.11.2",
-    "stylelint": "^13.3.2",
-    "stylelint-config-recommended": "^2.2.0",
     "stylelint-config-styled-components": "^0.1.1",
     "surge": "^0.21.3",
     "through2": "^3.0.1",
@@ -117,14 +115,19 @@
     "@devseed-ui/modal": "^1.0.3",
     "@fast-csv/format": "^4.3.5",
     "@turf/area": "^6.0.1",
-    "@turf/bbox": "^6.0.1",
+    "@turf/bbox": "^6.5.0",
     "@turf/bbox-polygon": "^6.0.1",
+    "@turf/boolean-point-in-polygon": "^6.5.0",
+    "@turf/circle": "^6.5.0",
     "@turf/combine": "^6.0.1",
+    "@turf/distance": "^6.5.0",
     "@turf/helpers": "^6.1.4",
     "@turf/square-grid": "^6.0.2",
     "@visx/legend": "^1.3.0",
     "@visx/scale": "^1.3.0",
     "blob-stream": "^0.1.3",
+    "brfs": "^2.0.2",
+    "chroma-js": "^2.4.2",
     "clipboard": "^2.0.6",
     "colormap": "^2.3.1",
     "d3": "^6.2.0",
@@ -156,6 +159,7 @@
     "react-helmet": "^6.1.0",
     "react-input-range": "^1.3.0",
     "react-joyride": "^2.2.1",
+    "react-papaparse": "^4.1.0",
     "react-redux": "^7.2.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
@@ -165,8 +169,10 @@
     "react-transition-group": "^4.4.1",
     "redux": "^4.0.5",
     "styled-components": "5.0.0",
+    "stylelint-config-standard": "^30.0.1",
     "stylelint-processor-styled-components": "^1.10.0",
     "topojson-client": "^3.1.0",
+    "turf-point": "^2.0.1",
     "utf8": "^3.0.0"
   }
 }
