@@ -19,6 +19,7 @@ import { multiply } from '../../styles/utils/math';
 import media from '../../styles/utils/media-queries';
 
 import LogoReverse from '../../../icons/collecticons/logo-reverse';
+import FeedbackForm from './feeback-form';
 
 const _rgba = stylizeFunction(rgba);
 
@@ -61,7 +62,6 @@ const PageNav = styled.nav`
   ${media.mediumUp`
     flex-flow: column nowrap;
   `}
-
 `;
 
 const GlobalMenu = styled.ul`
@@ -84,19 +84,20 @@ const GlobalMenu = styled.ul`
     height: 3rem;
     text-align: center;
   }
+  > *:nth-last-child(2) > * {
+    width: 3rem;
+    height: 3rem;
+    text-align: center;
+  }
 
   ${media.mediumUp`
     flex-flow: column nowrap;
     justify-content: center;
   `}
 
-
   ${media.mediumUp`
     > *:first-child {
       margin: 0;
-    }
-    > *:last-child {
-      margin-top: auto;
     }
   `}
 `;
@@ -153,7 +154,7 @@ const GlobalMenuLink = styled.a`
   &:visited {
     color: inherit;
   }
-  
+
   &:hover {
     opacity: 1;
     background: ${_rgba(themeVal('color.baseLight'), 0.08)};
@@ -176,7 +177,7 @@ const StyledNavLink = filterComponentProps(NavLink, propsToFilter);
 const StyledLink = filterComponentProps(Link, propsToFilter);
 
 class PageHeader extends React.Component {
-  render () {
+  render() {
     return (
       <PageHead role='banner'>
         <PageHeadInner>
@@ -219,6 +220,7 @@ class PageHeader extends React.Component {
               </li>
               <li>
                 <GlobalMenuLink
+                  id='info-box'
                   as={StyledNavLink}
                   exact
                   to='/about'
@@ -228,6 +230,9 @@ class PageHeader extends React.Component {
                 >
                   <span>About</span>
                 </GlobalMenuLink>
+              </li>
+              <li style={{ marginTop: 'auto', textAlign: 'center' }}>
+                <FeedbackForm />
               </li>
               <li>
                 <ShareOptions />
@@ -240,7 +245,6 @@ class PageHeader extends React.Component {
   }
 }
 
-PageHeader.propTypes = {
-};
+PageHeader.propTypes = {};
 
 export default PageHeader;
