@@ -1,9 +1,30 @@
 import { saveAs } from 'file-saver';
-import { getTimestamp, round } from '../../../utils/format';
+import { formatThousands, toTitleCase, getTimestamp, round } from '../../../utils/format';
 import config from '../../../config';
 const { indicatorsDecimals } = config;
+import { checkIncluded } from '../panel-data';
+import difference from 'lodash.difference';
 
-export default async function exportZonesGeoJSON (selectedArea, zones) {
+export default async function exportZonesGeoJSON (selectedArea, zones, selectedResource, filtersLists) {
+  let filters = Object.entries( filtersLists ).map( ([name, filter]) => {
+    if (filter.energy_type && !checkIncluded(filter, selectedResource)) {
+      return;
+    }
+    let title = filter.title || filter.name;
+    if (filter.unit) {
+      title = `${title} (${filter.unit})`;
+    }
+
+    let value = filter.input.value;
+    if (filter.isRange) {
+      return {...filter, min_value: value.min, max_value: value.max};
+    } else if ( typeof value === 'boolean' )
+    {
+      return {...filter, value: value};
+    }
+    return filter;
+  } ).filter( (x) => x );
+
   const featureCollection = {
     type: 'FeatureCollection',
     features: zones.map((z) => {
@@ -25,10 +46,11 @@ export default async function exportZonesGeoJSON (selectedArea, zones) {
             indicatorsDecimals.zone_output_density
           ),
           installed_capacity_potential_mw: summary.icp,
-          capacity_factor: round(summary.cf, indicatorsDecimals.cf)
+          capacity_factor: round(summary.cf, indicatorsDecimals.cf),
         }
       };
-    })
+    }),
+    filters_values: filters,
   };
 
   const blob = new Blob([JSON.stringify(featureCollection)], {
