@@ -2,5 +2,6 @@
 
 export default {
   environment: 'staging',
-  apiEndpoint: 'https://cenpsr6h9g.execute-api.us-east-2.amazonaws.com/v1'
+  apiEndpoint: 'https://cenpsr6h9g.execute-api.us-east-2.amazonaws.com/v1',
+  githubIssuesApi: 'https://cenpsr6h9g.execute-api.us-east-2.amazonaws.com/v1/feedback',
 };
